requirements.txt
Pillow
requests
selenium